test: Use a more functional syntax for geolocation mock
